update: completion for docker in mac

+++ bin/link.sh
@@ -8,11 +8,12 @@ for f in .??*
 do
     [[ "$f" == ".git" ]] && continue
     [[ "$f" == ".gitignore" ]] && continue
+    [[ "$f" == ".DS_Store" ]] && continue
     [[ "$f" == "Makefile" ]] && continue
     [[ "$f" == "README.md" ]] && continue
 
     [[ -L "$HOME/$f" ]] && unlink "$HOME/$f"
     [[ -f "$HOME/$f" ]] && mkdir -p ./backup && mv "$HOME/$f" ./backup/
-    ln -s "$SCRIPT_DIR/$f" "$HOME/$f"
+    ln -snfv "$SCRIPT_DIR/$f" "$HOME/$f"
     echo "$f"
 done
